Use correct type when comparing/looping over size types
Although conversions from unsigned int/longs to signed int should be safe since we're not working with vectors of several million elements which means overflow shouldn't happen, it is technically less safe, and more importantly, causes compiler warnings.

+++ src/control/positionControl/BBTrajectories/Trajectory1D.cpp
@@ -8,8 +8,8 @@
 
 namespace rtt {
 
-void Trajectory1D::addTrajectory(const std::vector<BB::BBTrajectoryPart> &newParts, double addFromTime) {
-    for (int i = 0; i < parts.size(); i++) {
+void Trajectory1D::addTrajectory(const std::vector<BB::BBTrajectoryPart>& newParts, double addFromTime) {
+    for (size_t i = 0; i < parts.size(); i++) {
         if (addFromTime <= parts[i].tEnd) {
             parts[i].tEnd = addFromTime;
             parts.erase(parts.begin() + 1 + i, parts.end());
